Add trackpad reading
Added reading from the trackpad for exact touch coordinates
This will be useful for the future when we actually use that to start
drawing to the screen.
I also have a terrible interface over it, but all will work out well
enough... probably...

For a quick rundown:
Buffer was a super quick heap buffer that lives on the stack. Do not
copy it around.
Notepad is the class that will be used to actually draw the notes to the
screen
Window is a simple wrapper around GLFW and ImGUI
Layer is an interface that lets us just create a class and have it draw
to the screen more easily
PaintLayer and NotepadLayer are the layers for the UI stuff
Device in the inputs are raw input device handling functions
Touchpad is a simple wrapper that gives quicker access to the touchpad
itself

I know the code is terrible, but it works ok... ish... I need to add
actual error checking, but that can happen in the future.